arm: dts: k3-j721e: Update OSPI settings for octal DTR mode
The Cypress Semper flash supports octal DTR mode, so update the device
tree settings to use 8 data lines for both Rx and Tx and allow using DTR
mode on both lines.

The read delay is set to 4. The calibration sequence yields a value of 2
but since the sequence is run before the flash is switched to octal DTR
mode, it gets that value for legacy mode. In octal DTR mode the flash
needs a read delay of 4. Setting read delay here over-rides the
calibrated value.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>